+++ src/Config/Hint/DefinitionHint.php
@@ -6,6 +6,7 @@ namespace WoohooLabs\Zen\Config\Hint;
 use WoohooLabs\Zen\Container\Definition\ClassDefinition;
 use WoohooLabs\Zen\Container\Definition\DefinitionInterface;
 use WoohooLabs\Zen\Container\Definition\ReferenceDefinition;
+use WoohooLabs\Zen\Exception\ContainerException;
 
 class DefinitionHint extends AbstractHint implements DefinitionHintInterface
 {
@@ -13,6 +14,16 @@ class DefinitionHint extends AbstractHint implements DefinitionHintInterface
      * @var string
      */
     private $className;
+
+    /**
+     * @var array
+     */
+    private $parameters;
+
+    /**
+     * @var array
+     */
+    private $properties;
 
     public static function singleton(string $className): DefinitionHint
     {
@@ -28,6 +39,36 @@ class DefinitionHint extends AbstractHint implements DefinitionHintInterface
     {
         parent::__construct($scope);
         $this->className = $className;
+        $this->parameters = [];
+        $this->properties = [];
+    }
+
+    /**
+     * @param string|int|float|bool|array|null $value
+     */
+    public function parameter(string $name, $value): DefinitionHint
+    {
+        if (is_scalar($value) === false && is_array($value) === false) {
+            throw new ContainerException("Constructor argument '$name' in '$this->className' must be a scalar or an array!");
+        }
+
+        $this->parameters[$name] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string|int|float|bool|array|null $value
+     */
+    public function property(string $name, $value): DefinitionHint
+    {
+        if (is_scalar($value) === false && is_array($value) === false) {
+            throw new ContainerException("Property '$this->className::\$$name' must be a scalar or an array!");
+        }
+
+        $this->properties[$name] = $value;
+
+        return $this;
     }
 
     /**
@@ -39,7 +80,13 @@ class DefinitionHint extends AbstractHint implements DefinitionHintInterface
     {
         if ($this->className === $id) {
             return [
-                $id => new ClassDefinition($this->className, $this->getScope(), $isAutoloaded),
+                $id => new ClassDefinition(
+                    $this->className,
+                    $this->getScope(),
+                    $isAutoloaded,
+                    $this->parameters,
+                    $this->properties
+                ),
             ];
         }
 
@@ -58,7 +105,6 @@ class DefinitionHint extends AbstractHint implements DefinitionHintInterface
 
         return $result;
     }
-
     /**
      * @internal
      */